Successfully refactored for Webpack
