+++ package.json
@@ -1,21 +1,37 @@
 {
   "name": "nathansimpson.design",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
+  "description": "Another go at the nathansimpson.design site, without Create React App",
+  "main": "index.js",
+  "scripts": {
+    "test": "react-scripts test --env=jsdom",
+    "start": "npm run build",
+    "build": "webpack --d && webpack-dev-server --d --content-base public/ --inline --hot --port 1234"
+  },
+  "author": "Nathan Simpson",
+  "license": "ISC",
   "dependencies": {
     "@fortawesome/fontawesome": "^1.1.3",
     "@fortawesome/fontawesome-free-brands": "^5.0.6",
     "@fortawesome/fontawesome-free-regular": "^5.0.6",
     "@fortawesome/react-fontawesome": "0.0.17",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.4",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
+    "react-markdown": "^3.3.0",
     "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.0"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+  "devDependencies": {
+    "node-sass": "^4.7.2",
+    "sass-loader": "^6.0.7",
+    "svg-inline-loader": "^0.8.0",
+    "webpack": "^4.1.1",
+    "webpack-cli": "^2.0.12",
+    "webpack-dev-server": "^3.1.1"
   }
 }
