+++ src/components/Portfolio.js
@@ -22,7 +22,7 @@ class PortfolioFilter extends React.Component {
 class PortfolioItem extends React.Component {
  render(){
    return(
-     <div className="portfolio_item">
+     <div className="portfolio_item" key={this.props.data.key}>
        <img className="portfolio_item_image" src={"http://files.nathansimpson.design/portfolio/" + this.props.data.imagesrc} alt=""/>
         <span className="portfolio_item_label">{this.props.data.projectName}</span>
         <div className="portfolio_item_gradient"></div>
