+++ src/index.html
@@ -1,0 +1,13 @@
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>Nathan Simpson</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <link rel="shortcut icon" href="favicon.ico">
+
+  </head>
+  <body>
+    <div id="app" />
+    <script src="../public/bundle.js" type="text/javascript"></script>
+  </body>
+</html>
